moved trackUtils functions to innerSliderutils
