+++ src/inner-slider.js
@@ -220,7 +220,6 @@ export class InnerSlider extends React.Component {
     this.setState(state)
     if (triggerSlideHandler === undefined) return
     this.slideHandler(triggerSlideHandler)
-    console.log('handling slide to:', triggerSlideHandler)
   }
   slickPrev = () => {
     // this and fellow methods are wrapped in setTimeout
